+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v0.5.0
+
+### Features
+
+- [python][venv-create] Allow to specify a custom Python interpreter when creating virtual env with `--interpreter`
+- Add `venv-activate` module to `python` collection
+- Add `cache-clear` module to `python` collection
+
+
 ## v0.4.0
 
 ### Features